aks-desktop: Split up large a11y test file
